refactor: extract verified-email reset flow to reduce PR diff footprint
